powerpc/sysdev: Add contutto driver
The Con Tutto card is a replacement for the Custom DIMMs used on IBM
manufactured Power 8 systems. The card contains an FPGA which
communicates with the P8 over the high-speed Differential Memory
Interface used the Centaur memory buffer. This setup allows the card to
host system memory and serve as a platform for near-memory accerators
and other tightly coupled applications.

The memory on the card is exposed to userspace through libnvdimm which
allows the user use it as a pmem or DAX device, depending on their
application.

Not-signed-off-by: Oliver O'Halloran <oohall@gmail.com>