HACK: nvdimm: loads rather than stores for WPQ flush
This replaces the WPQ flush trigger store with a load. This is required
due to the old contutto images using a load rather than a store
as a flush trigger.

This patch also replaces wmb() with full barriers since cache flushes
are treated as loads for ordering purposes on ppc.

Not-signed-off-by: Oliver O'Halloran <oohall@gmail.com>