drivers/nvdimm: Use HPAGE_PMD_SIZE over HPAGE_SIZE
Inside of libnvdimm HPAGE_SIZE is used to calculate how large the altmap
region needs to be to support the number of PFNs in a region and any
padding required to align the region to the vmemmap page size. On PPC
using HPAGE_SIZE breaks the build because it depends on an unexported
symbol (HPAGE_SHIFT) which is set at boot time. The obvious fix is to
just export it, but I'm not convinced that is a good fix since
HPAGE_SIZE and the vmemmap mapping size aren't necessarily the same.

As an interim, use HPAGE_PMD_SIZE since that's the seems to be the
same everywhere.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>