mm/huge_memory: don't open code zap_deposited_table()
Depending flags of the PMD being zapped there may or may not be a deposited
pgtable to be freed. In two of the three cases this actual freeing is open
coded while the third uses the helper function zap_deposited_table(). This
converts the other usages to use the helper.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
---
For reference this is the contents of zap_deposited_table()

        pgtable_t pgtable;

        pgtable = pgtable_trans_huge_withdraw(mm, pmd);
        pte_free(mm, pgtable);
        atomic_long_dec(&mm->nr_ptes);