mm/memblock: make memblock_alloc_base_nid non-static
Doesn't seem to be a whole lot of reasons to keep this static.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>

---
TODO: this is the wrong API for this. Also memblock allocations are done
bottom-up by default so we should tbe careful about testing this. We
might need to hack it up a bit to ensure that highmem allocations work.