powerpc/pseries: Rework reserved hugepage handling
On pseries systems the hypervisor needs to specially reserve contigious
memory so that 16GB gigantic pages can be used by the guest. These
contigious ranges are exposed to the guest as a memory@<addr> device
tree node with the ibm,expected#pages property which indicates the
number of huge pages the guest can create from that region.

This patch converts the existing hash-only handling of these huge page
regions to be MMU agnostic and converts it to use the new pre-boot
huge page donation API rather than the add_gpage() hack. This has the
happy side effect of fixing an overflow when more than 16TB of reserved
memory is given to the guest.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>