mm/hugetlb: Fix use after free gather_bootmem_prealloc()
While traversing the list of hugepages allocated at boot and adding them
to their respective hstate we free the current list entry in the HIGHMEM
case. This bug actually causing a crash would require some alignment of
the stars, but it's still a bit sketch. This patch fixes it by using the
_safe iterator.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>