cpuidle/powernv: Respect the nap enabled flag
The "kernel.powersave-nap" sysctl flag is supposed to be used for
disabling the usage of the nap power saving state. This is sometimes
useful to stop the core from entering low-power modes which prevent
debug access to it's internal registers.

There are several methods for invoking the nap instruction in the
kernel: power7_idle(), power7_idle_type() and power7_idle_instr().
the _instr() variant is a low-level primitive used for various HW
re-init sequences so it should not check the flag, but
power7_idle_type() should.

One thing to note is that setting powersave=off in the kernel command
line is *not* a substutite for disabling nap mode. Disabling powersaving
entirely prevents SMT mode switching because the unused threads can
never be fully disabled.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>