powerpc/eeh: Split eeh_probe into probe_pdn and probe_pdev
The EEH core has a concept of "early probe" and "late probe." When the
EEH_PROBE_MODE_DEVTREE flag is set (i.e pseries) we call the eeh_ops->probe()
function in eeh_add_device_early() so the eeh_dev state is initialised based on
the pci_dn. It's important to realise that this happens *long* before the PCI
device has been probed and a pci_dev structure created. This is necessary due
to a PAPR requirement that EEH be enabled before to OS starts interacting
with the device.

The late probe done in eeh_add_device_late() when the EEH_PROBE_MODE_DEV flag
is set (i.e. PowerNV). The main difference is that this happens *after* the
pci_dev has been created. As a result there is no actual dependency on a pci_dn
in the late probe case. Splitting the single eeh_ops->probe() function into
seperate functions allows us to simplify the late probe case. This patch
does that split in preperation for refactoring the late probe method.

It also changes the prototype for the probe functions to be void. Currently
they return a void *, but both implementations always return NULL.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>