powerpc/eeh: Rework the late probe
This adjusts how the EEH late probe is used a bit to remove the use of
pci_dn in the EEH core. To do this we move the code to map a pci_dev to
it's corresponding eeh_dev into platform specific code rather than
having the generic EEH code looking at pci_dn directly.

For pseries the pci_dn lookup still makes sense so move that into
a pseries_eeh_probe_pdev() function.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>