powerpc/eeh: Remove EEH_PE_BUS flag
This flag is set, but never looked at by anything. Removing it also allows us
to clean up eeh_add_to_parent_pe() a bit since the only difference in behaviour
now is that we clear the EEH_PE_INVALID flag if it's set on an existing PE.
Clearing INVALID is harmless if it's already clear, so fold the two cases
together.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>