powerpc/eeh: Delay freeing eeh_pe's until recovery is done - FIXME: clear CLEANUP if we add a new device?
When the last device in an eeh_pe is removed the eeh_pe structure itself (and
any empty parents) are freed since they are no longer needed. This results in a
crash when a hotplug driver is involved since the following may occur:

1. Device is suprise removed.
2. Driver performs an MMIO, which fails and causes an eeh_event to be queued.
3. Hotplug driver recieves a hotplug interrupt and removes pci_devs the slot.
4. pci_dev is torn down and the eeh_pe is freed.
5. The EEH event handler thread processes the eeh_event and crashes since the
   eeh_pe pointer in the eeh_event structure is no longer valid.

Crashing is generally considered poor form so instead, mark the PE as
undergoing recovery before it's inserted into the event queue. While
the EEH_PE_RECOVERING flag is set eeh_rmv_from_parent_pe() will set the
EEH_PE_CLEANUP flag instead of freeing the eeh_pe. Once the event handler
has finished recovering the PE it will prune any empty PEs left.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
---
Sam Bobroff is working on implemented proper refcounting for eeh_pe's,
but that's not going to be ready for a while and this is is less broken
than what we have now.