powerpc/eeh: Defer printing stack trace
Currently we print a stack trace in the event handler to help with debugging
EEH issues. In the case of suprise hot-unplug this is unneeded, so we want to
prevent printing the stack trace unless we know it's due to an actual device
error. To accomplish this, we can save a stack trace at the point of detection
and only print it once the EEH recovery handler has determined the freeze was
due to an actual error.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>