powerpc/pcidn: Rename add/remove_pci_dev_data so something sriov specific
Much of the powerpc PCI code assumes that a pci_dn structure exists for all
devices in the system. This is (mostly) fine for real devices since we get a
node in the device-tree from either firmware, or from a hotplug driver. For
SR-IOV virtual functions we need to manually create the pci_dn structures
which is handled by add_pci_dev_data and remove_pci_dev_data. These functions
are *only* used to instantiate pci_dn's for VFs so rename them to something
SR-IOV specific rather than having them masquerade as generic code.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>