powerpc/sriov: Fix recovery for un-aware SR-IOV PFs
Enabling SR-IOV requires cooperation from the driver that is bound to the
physical function since the exact mechanics of managing VFs is device
dependent. As a result, if the PF driver does not implement the PCI error
handling callbacks then we have no way to restore the SR-IOV state after
resetting the device and we must remove all the VFs from the system when
an error occurs.

Currently we remove the pci_dev's for each virtual function, but we do not
tear down the corresponding EEH state for the VF. This results in a
use-after-free later in the recovery process since the eeh_dev for the
VF is still part of the eeh_pe's child device list. This patch fixes
this by ensuring that the EEH state is removed when before the pci_dn
for the VF is removed. This also has the effect of correctly cleaning up
the eeh_pe that contains the VF.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>