powerpc/powernv: Allow non-prefetch BARs in the 64bit window on PHB3/4
The PowerNV SR-IOV support requires VF BARs to be placed in the 64-bit
bridge window since the PHB driver relies on being able to use the M64BT
to configure the mapping of MMIO addresses to EEH PE numbers. As the name
suggests the M64BT is only usable for the 64bit window so SR-IOV devices
with non-prefetchable BARs cause some headaches.

The Linux BAR allocator places all non-prefetchable BARs, including 64bit BARs,
into the non-prefetchable (32bit) bridge window by default since that's the
safest option. This is however not required. The PCIe Base Spec does not
*require* this behaviour and provides some guidence on when it's permitted to
place a non-prefetchable BAR in the prefetchable window. From PCIe Base Spec 5.0-PUB
pg 699:
Section 7.5.1.2.1 - Base Address Registers:

1) The entire path from the host to the adapter is over PCI Express.
2) No conventional PCI or PCI-X devices do peer-to-peer reads to the range
   mapped by the BAR.

3) The PCI Express Host Bridge does no byte merging. (This is believed to be
   true on most platforms.)
4) Any locations with read side-effects are never the target of Memory Reads
   with the TH bit Set. See Section 2.2.5.
5) The range mapped by the BAR is never the target of a speculative Memory Read,
   either Host initiated or peer-to-peer.

The first two conditions are obviously true if

If we know the PCI heirachy is entirely PCIe then 1) and 2) and trivially
true. 4) Is

3) and 5) are dependent on the behaviour of the root complex.

Not-Signed-off-by: Oliver O'Halloran <oohall@gmail.com>