powernv/eeh: Use generic code to handle hot resets
When we reset PCI devices managed by a hotplug driver the reset may generate
spurious hotplug events that cause the PCI device we're resetting to be torn
down accidently. This is a problem for EEH (when the driver is EEH aware)
since we want to leave the OS PCI device state intact so that the device
can be re-set without losing any resources (network, disks, etc) provided
by the driver.

Generic PCI code provides the pci_bus_error_reset() function to handle
resetting a PCI Device (or bus) by using the reset method provided by
the hotplug slot driver. We can use this function if the EEH core has
requested a hot reset (common case) without tripping over the hotplug
driver.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
---
I know that include is a bit gross, but:

a) We're already doing it in pci-ioda.c, and in pseries/pci.
b) It's pci_bus_error_reset() isn't really a function that
   should be provided to non-pci core code.