selftests/powerpc: Check the driver rebinds when testing EEH
While working on EEH I have noticed that occasionally drivers will not re-bind
after an EEH recovery cycle. This is usually due to the EEH PE not being
correctly unfrozen or some part of the device state not being restored. To
catch these situations verify that the same driver is bound to the device
before and after the recovery pass.

This is mainly an issue for EEH un-aware devices where the device is removed
and then re-scanned.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>