powerpc/eeh: Remove eeh_dev_phb_init_dynamic()
This function is a one line wrapper around eeh_phb_pe_create() and despite the name
it doesn't create any eeh_dev structures. Replace it with direct calls to
eeh_phb_pe_create() since it's one less layer of pointless indirection and
it actually does what it says on the tin.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>