powernv/pci: Move IODA PE configuration to pcibios_bus_add_device()
Currently we do this right after the bridge windows are assigned in
pcibios_setup_bridge(). This makes a degree of sense since for conventional
PHBs the IODA PE is generally determined by the bridge window which contains
the devices MMIO BARs. However, it results in a strange asymmetry in when
the pe->device_count field is incremented and decremented.

When the IODA PE is first configured we walk the devices under the bridge
and set the pdn->pe_number field as needed. The device_count for the PE is
only incremented if the the pe_number was not previously set, so we don't
double-count devices. The device_count is decremented when
pcibios_device_release() is called, i.e. just before it's free()ed.

This scheme works because when a new PCI device is scanned we

we currently traverse the bus and for
each device with an unset pdn->pe_number we bump the device_count for the
PE. The device_count is decremented when pcibios_device_release() is called
which creates a bit of an asymmetry in the device_count accounting since
it's not entirely clear the pcibios_setup_bridge() will be called after
the first scan.

Move the configuration of the IODA PE into pcibios_bus_add_device() since
that's called for each device on a bus after the MMIO BARs for the device
have been assigned. This allows us to delete the

Cc: Sergey Miroshnichenko <s.miroshnichenko@yadro.com>
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>