powerpc/eeh: Add sysfs files in late probe
Move creating the EEH specific sysfs files into the eeh_add_device_late().
This is done immediately after eeh_add_device_late() in most cases anyway
so fold it in.

There's two cases wwhere this isn't immediately after the late probe, the
first being the OF based PCI probing code and the second being for VFs on
pseries. The first case it's fine since the OF probe and the DT traversal
will visit devices in the same order. The second case is a little more
interesting.

For whatever reason the eeh_pe that a VF is added to is wrong and needs to
be fixed up. Currently this is done after running the late probe, but it
doesn't need to be since nothing in the late probe step looks at the eeh_pe
so we can move the fixup to before the late probe without issues.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>