powerpc/eeh: Moving finding the parent PE into the platform
Currently the generic EEH code uses the pci_dn of a device to look up the
PE of the device's parent bridge, or physical function. We don't want
PowerNV to be dependent on a pci_dn existing for a device so re-work
eeh_add_to_parent_pe() so that the platform determines what the parent
PE should be.

This allows us to use the scanned PCI topology to work out PE
relationships rather than the pci_dn on PowerNV. Pseries requires
the pci_dn for other reasons so move the existing lookup to the pseries
containment zone.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
---
"parent" meaning "parent of the PE that actually contains this edev"
is stupid, but it's stupid consistent with what's there already. Also
I couldn't think of a way to fix it without adding a bunch of boring
boilerplate at the call sites.