powerpc/pci-hotplug: Scan the whole bus when using PCI_PROBE_NORMAL
Currently when using the normal (i.e not building pci_dev's from the DT node)
we only scan the device corresponding to the first child DT node under the
bridge's DT node. This doesn't make much sense to me, but it seems to have
worked so far, so at a guess it works in a PCIe environment where there is
typically only a single downstream device at devfn 00.0

In any case it's completely broken when no pci_dn is available. Remove
the PCI_DN checking and scan each of the device number that might be on
the downstream bus.

Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
---
We should probably be using pci_scan_child_bus() here, but I'm sure
there's some insane legacy reason we're using pci_scan_slot()
directly. Urgh...