powernv/iov: Ensure the pdn for VFs always contains a valid PE number
On PowerNV we use the pcibios_sriov_enable() hook to do two things:

1. Create a pci_dn structure for each of the VFs, and
2. Configure the PHB's internal BARs that map MMIO ranges to PEs
   so that each VF has its own PE. Note that the PE also determines
   the IOMMU table the HW uses for the device.

Currently we do not set the pe_number field of the pci_dn immediately after
assigning the PE number for the VF that it represents. Instead, we do that
in a fixup (see pnv_pci_dma_dev_setup) which is run inside the
pcibios_add_device() hook which is run prior to adding the device to the
bus.

On PowerNV we add the device to its IOMMU group using a bus notifier and
in order for this to work the PE number needs to be known when the bus
notifier is run. Adding the VF to an IOMMU group at this point work today
since the fixup ensures that the PE number is known when the notifier runs.

We would like to move the call to pcibios_setup_device(), which runs the fixup,
out of pcibios_add_device() and into pcibios_bus_add_device(). However, this
would result in the fixup occuring long after the bus notifier which breaks the
IOMMU group setup.

We can fix this by moving the initialisation of pdn->pe_number out of the fixup
and into pnv_pci_ioda_setup_vf_PE() where the hardware PE state is configured.
This move causes two problems though:

1. We trip the WARN_ON() in the fixup code, and
2. The EEH core will clear pdn->pe_number while recovering VFs.

The only justification for either of these is a comment in eeh_rmv_device()
suggesting that pdn->pe_number *must* be set to IODA_INVALID_PE in order
for the VF to be scanned. However, this comment appears to have no basis in
reality so just delete it and the corresponding code.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
---
v2: Grammar fixes, reworded commit message slightly.
---
Can't get rid of the fixup entirely since we need it to set the
ioda_pe->pdev back-pointer. I'll look at killing that another time.